Move algorithm maps to header
