+++ source/llassetgen-cmd/CMakeLists.txt
@@ -23,6 +23,7 @@ set(source_path  "${CMAKE_CURRENT_SOURCE_DIR}/source")
 
 set(headers
     ${include_path}/CLI11.h
+    ${include_path}/algorithms.h
     ${include_path}/helpstrings.h)
 
 set(sources
